vendor-service/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
gunicorn
requests
pika
boto3
django-cors-headers
drf-yasg
celery==5.6.0
redis==4.5.5
